src/backend/requirements.txt
asgiref==3.8.1
astroid==3.3.8
black==24.10.0
click==8.1.8
dill==0.3.9
Django==5.1.6
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pylint==3.3.3
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
